+++ docs/hsc/process/sidebar-items.js
@@ -1,1 +1,0 @@
-window.SIDEBAR_ITEMS = {"enum":["SavingCells","Stats2Save"],"struct":["CellDivisionProbabilities","Exponential","Moran","ProcessOptions","SavingOptions","Snapshot"]};