+++ docs/trait.impl/std/io/trait.BufRead.js
@@ -1,3 +1,0 @@
-(function() {var implementors = {
-"indicatif":[["impl&lt;R: <a class=\"trait\" href=\"https://doc.rust-lang.org/1.75.0/std/io/trait.BufRead.html\" title=\"trait std::io::BufRead\">BufRead</a>&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/1.75.0/std/io/trait.BufRead.html\" title=\"trait std::io::BufRead\">BufRead</a> for <a class=\"struct\" href=\"indicatif/struct.ProgressBarIter.html\" title=\"struct indicatif::ProgressBarIter\">ProgressBarIter</a>&lt;R&gt;"]]
-};if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()